refactor: split repository method families
