+++ polylogue/storage/repository.py
@@ -10,45 +10,30 @@ fetching of conversations, messages, and attachments together.
 
 from __future__ import annotations
 
-import asyncio
 import builtins
-from collections.abc import AsyncIterator
 from typing import TYPE_CHECKING
 
-from polylogue.lib.models import Conversation, ConversationSummary, Message
-from polylogue.logging import get_logger
+from polylogue.protocols import ConversationReader, SearchStore, TagStore
 from polylogue.storage.backends.async_sqlite import SQLiteBackend
-from polylogue.storage.embedding_stats import read_embedding_stats_async
-from polylogue.storage.hydrators import (
-    conversation_from_records,
-    conversation_summary_from_record,
-    message_from_record,
-)
-from polylogue.storage.search_cache import invalidate_search_cache
-from polylogue.storage.store import (
-    AttachmentRecord,
-    ContentBlockRecord,
-    ConversationRecord,
-    ConversationRenderProjection,
-    MessageRecord,
-    PublicationRecord,
-    RunRecord,
-)
-from polylogue.types import ConversationId
 
 if TYPE_CHECKING:
     from pathlib import Path
 
-    from polylogue.lib import filters
-    from polylogue.lib.stats import ArchiveStats
     from polylogue.storage.backends.query_store import SQLiteQueryStore
 
-from polylogue.protocols import ConversationReader, SearchStore, TagStore, VectorProvider
+from polylogue.storage.repository_reads import RepositoryReadMixin
+from polylogue.storage.repository_vectors import RepositoryVectorMixin
+from polylogue.storage.repository_writes import RepositoryWriteMixin
 
-logger = get_logger(__name__)
 
-
-class ConversationRepository(ConversationReader, SearchStore, TagStore):
+class ConversationRepository(
+    RepositoryReadMixin,
+    RepositoryWriteMixin,
+    RepositoryVectorMixin,
+    ConversationReader,
+    SearchStore,
+    TagStore,
+):
     """Async repository for conversation storage operations.
 
     Wraps SQLiteBackend to provide high-level async storage interface with
@@ -141,1055 +126,6 @@ class ConversationRepository(ConversationReader, SearchStore, TagStore):
     async def close(self) -> None:
         """Close database connections and release resources."""
         await self._backend.close()
-
-    # --- Read Methods ---
-
-    async def resolve_id(self, id_prefix: str) -> ConversationId | None:
-        """Resolve a partial ID to a full conversation ID.
-
-        Args:
-            id_prefix: Full or partial conversation ID
-
-        Returns:
-            Full ConversationId if resolved, None if ambiguous or not found
-        """
-        resolved = await self.queries.resolve_id(id_prefix)
-        return ConversationId(resolved) if resolved else None
-
-    async def get(self, conversation_id: str) -> Conversation | None:
-        """Get a conversation with all messages and attachments (eager loading).
-
-        Fetches conversation, messages, and attachments in parallel.
-
-        Args:
-            conversation_id: Full conversation ID
-
-        Returns:
-            Conversation with all data loaded, or None if not found
-        """
-        conv_record = await self.queries.get_conversation(conversation_id)
-        if not conv_record:
-            return None
-
-        # Fetch messages and attachments in parallel
-        msg_records, att_records = await asyncio.gather(
-            self.queries.get_messages(conversation_id),
-            self._backend.get_attachments(conversation_id),
-        )
-
-        return conversation_from_records(conv_record, msg_records, att_records)
-
-    async def get_render_projection(self, conversation_id: str) -> ConversationRenderProjection | None:
-        """Fetch repository-owned render projection with raw attachment layout preserved."""
-        conv_record = await self.queries.get_conversation(conversation_id)
-        if not conv_record:
-            return None
-
-        msg_records, att_records = await asyncio.gather(
-            self.queries.get_messages(conversation_id),
-            self._backend.get_attachments(conversation_id),
-        )
-        return ConversationRenderProjection(
-            conversation=conv_record,
-            messages=msg_records,
-            attachments=att_records,
-        )
-
-    async def view(self, conversation_id: str) -> Conversation | None:
-        """Get a conversation with ID resolution support.
-
-        Attempts to resolve a partial ID to a full ID, then fetches the conversation.
-
-        Args:
-            conversation_id: Full or partial conversation ID
-
-        Returns:
-            Conversation with all data loaded, or None if not found
-        """
-        full_id = await self.resolve_id(conversation_id) or conversation_id
-        return await self.get(str(full_id))
-
-    async def get_eager(self, conversation_id: str) -> Conversation | None:
-        """Get a conversation with all messages and attachments.
-
-        Alias for get() - all async operations are eager by design.
-
-        Args:
-            conversation_id: Full conversation ID
-
-        Returns:
-            Conversation with all data loaded, or None if not found
-        """
-        return await self.get(conversation_id)
-
-    async def _hydrate_conversations(
-        self,
-        conversation_records: builtins.list[ConversationRecord],
-        *,
-        ordered_ids: builtins.list[str] | None = None,
-    ) -> builtins.list[Conversation]:
-        """Hydrate eager conversation models from already-fetched records."""
-        if not conversation_records:
-            return []
-
-        by_id = {record.conversation_id: record for record in conversation_records}
-        conversation_ids = ordered_ids or [record.conversation_id for record in conversation_records]
-        present_ids = [conversation_id for conversation_id in conversation_ids if conversation_id in by_id]
-        if not present_ids:
-            return []
-
-        msgs_by_id, atts_by_id = await asyncio.gather(
-            self.queries.get_messages_batch(present_ids),
-            self._backend.get_attachments_batch(present_ids),
-        )
-        return [
-            conversation_from_records(
-                by_id[conversation_id],
-                msgs_by_id.get(conversation_id, []),
-                atts_by_id.get(conversation_id, []),
-            )
-            for conversation_id in present_ids
-        ]
-
-    async def get_summary(self, conversation_id: str) -> ConversationSummary | None:
-        """Get a conversation summary without loading messages.
-
-        Args:
-            conversation_id: Full conversation ID
-
-        Returns:
-            ConversationSummary, or None if not found
-        """
-        conv_record = await self.queries.get_conversation(conversation_id)
-        if not conv_record:
-            return None
-        return conversation_summary_from_record(conv_record)
-
-    async def list_summaries(
-        self,
-        limit: int | None = 50,
-        offset: int = 0,
-        provider: str | None = None,
-        providers: builtins.list[str] | None = None,
-        source: str | None = None,
-        since: str | None = None,
-        until: str | None = None,
-        title_contains: str | None = None,
-        has_tool_use: bool = False,
-        has_thinking: bool = False,
-        min_messages: int | None = None,
-        max_messages: int | None = None,
-        min_words: int | None = None,
-        has_file_ops: bool = False,
-        has_git_ops: bool = False,
-        has_subagent: bool = False,
-    ) -> builtins.list[ConversationSummary]:
-        """List conversation summaries without loading messages.
-
-        Args:
-            limit: Maximum number of results
-            offset: Number of results to skip
-            provider: Filter by single provider name
-            providers: Filter by multiple providers
-            source: Filter by source name
-            since: Filter to conversations updated on/after this ISO date
-            until: Filter to conversations updated on/before this ISO date
-            title_contains: Filter by title substring (case-insensitive)
-            has_tool_use: Only conversations with tool_use blocks
-            has_thinking: Only conversations with thinking blocks
-            min_messages: Minimum message count
-            max_messages: Maximum message count
-            min_words: Minimum total word count
-            has_file_ops: Only conversations with file operations
-            has_git_ops: Only conversations with git operations
-            has_subagent: Only conversations with subagent blocks
-
-        Returns:
-            List of ConversationSummary objects
-        """
-        conv_records = await self.queries.list_conversations(
-            limit=limit,
-            offset=offset,
-            **self._conversation_filter_kwargs(
-                source=source,
-                provider=provider,
-                providers=providers,
-                since=since,
-                until=until,
-                title_contains=title_contains,
-                has_tool_use=has_tool_use,
-                has_thinking=has_thinking,
-                min_messages=min_messages,
-                max_messages=max_messages,
-                min_words=min_words,
-                has_file_ops=has_file_ops,
-                has_git_ops=has_git_ops,
-                has_subagent=has_subagent,
-            ),
-        )
-        return [conversation_summary_from_record(rec) for rec in conv_records]
-
-    async def iter_summary_pages(
-        self,
-        *,
-        page_size: int = 50,
-        provider: str | None = None,
-        providers: builtins.list[str] | None = None,
-        source: str | None = None,
-        since: str | None = None,
-        until: str | None = None,
-        title_contains: str | None = None,
-        has_tool_use: bool = False,
-        has_thinking: bool = False,
-        min_messages: int | None = None,
-        max_messages: int | None = None,
-        min_words: int | None = None,
-        has_file_ops: bool = False,
-        has_git_ops: bool = False,
-        has_subagent: bool = False,
-    ) -> AsyncIterator[builtins.list[ConversationSummary]]:
-        """Iterate summary pages in backend sort order without materializing all results."""
-        offset = 0
-        while True:
-            page = await self.list_summaries(
-                limit=page_size,
-                offset=offset,
-                provider=provider,
-                providers=providers,
-                source=source,
-                since=since,
-                until=until,
-                title_contains=title_contains,
-                has_tool_use=has_tool_use,
-                has_thinking=has_thinking,
-                min_messages=min_messages,
-                max_messages=max_messages,
-                min_words=min_words,
-                has_file_ops=has_file_ops,
-                has_git_ops=has_git_ops,
-                has_subagent=has_subagent,
-            )
-            if not page:
-                break
-            yield page
-            if len(page) < page_size:
-                break
-            offset += len(page)
-
-    async def list(
-        self,
-        limit: int | None = 50,
-        offset: int = 0,
-        provider: str | None = None,
-        providers: builtins.list[str] | None = None,
-        since: str | None = None,
-        until: str | None = None,
-        title_contains: str | None = None,
-        has_tool_use: bool = False,
-        has_thinking: bool = False,
-        min_messages: int | None = None,
-        max_messages: int | None = None,
-        min_words: int | None = None,
-        has_file_ops: bool = False,
-        has_git_ops: bool = False,
-        has_subagent: bool = False,
-    ) -> builtins.list[Conversation]:
-        """List conversations with eager-loaded messages and attachments.
-
-        Args:
-            limit: Maximum number of results
-            offset: Number of results to skip
-            provider: Filter by single provider name
-            providers: Filter by multiple providers
-            since: Filter to conversations updated on/after this ISO date
-            until: Filter to conversations updated on/before this ISO date
-            title_contains: Filter by title substring (case-insensitive)
-            has_tool_use: Only conversations with tool_use blocks
-            has_thinking: Only conversations with thinking blocks
-            min_messages: Minimum message count
-            max_messages: Maximum message count
-            min_words: Minimum total word count
-            has_file_ops: Only conversations with file operations
-            has_git_ops: Only conversations with git operations
-            has_subagent: Only conversations with subagent blocks
-
-        Returns:
-            List of Conversation objects with all data eager-loaded
-        """
-        conv_records = await self.queries.list_conversations(
-            limit=limit,
-            offset=offset,
-            **self._conversation_filter_kwargs(
-                provider=provider,
-                providers=providers,
-                since=since,
-                until=until,
-                title_contains=title_contains,
-                has_tool_use=has_tool_use,
-                has_thinking=has_thinking,
-                min_messages=min_messages,
-                max_messages=max_messages,
-                min_words=min_words,
-                has_file_ops=has_file_ops,
-                has_git_ops=has_git_ops,
-                has_subagent=has_subagent,
-            ),
-        )
-
-        return await self._hydrate_conversations(conv_records)
-
-    async def count(
-        self,
-        provider: str | None = None,
-        providers: builtins.list[str] | None = None,
-        since: str | None = None,
-        until: str | None = None,
-        title_contains: str | None = None,
-        has_tool_use: bool = False,
-        has_thinking: bool = False,
-        min_messages: int | None = None,
-        max_messages: int | None = None,
-        min_words: int | None = None,
-        has_file_ops: bool = False,
-        has_git_ops: bool = False,
-        has_subagent: bool = False,
-    ) -> int:
-        """Count conversations matching filters.
-
-        Args:
-            provider: Filter by single provider
-            providers: Filter by multiple providers
-            since: Filter to conversations updated on/after this ISO date
-            until: Filter to conversations updated on/before this ISO date
-            title_contains: Filter by title substring
-            has_tool_use: Only conversations with tool_use blocks
-            has_thinking: Only conversations with thinking blocks
-            min_messages: Minimum message count
-            max_messages: Maximum message count
-            min_words: Minimum total word count
-            has_file_ops: Only conversations with file operations
-            has_git_ops: Only conversations with git operations
-            has_subagent: Only conversations with subagent blocks
-
-        Returns:
-            Count of matching conversations
-        """
-        filters = self._conversation_filter_kwargs(
-            provider=provider,
-            providers=providers,
-            since=since,
-            until=until,
-            title_contains=title_contains,
-            has_tool_use=has_tool_use,
-            has_thinking=has_thinking,
-            min_messages=min_messages,
-            max_messages=max_messages,
-            min_words=min_words,
-            has_file_ops=has_file_ops,
-            has_git_ops=has_git_ops,
-            has_subagent=has_subagent,
-        )
-        filters.pop("source")
-        return await self.queries.count_conversations(**filters)
-
-    async def get_parent(self, conversation_id: str) -> Conversation | None:
-        """Get the parent conversation if it exists.
-
-        Args:
-            conversation_id: Child conversation ID
-
-        Returns:
-            Parent Conversation, or None if no parent
-        """
-        conv_record = await self.queries.get_conversation(conversation_id)
-        if conv_record and conv_record.parent_conversation_id:
-            return await self.get(str(conv_record.parent_conversation_id))
-        return None
-
-    async def get_children(self, conversation_id: str) -> builtins.list[Conversation]:
-        """Get all direct children of this conversation.
-
-        Args:
-            conversation_id: Parent conversation ID
-
-        Returns:
-            List of child Conversation objects with all data eager-loaded
-        """
-        child_records = await self.queries.list_conversations(parent_id=conversation_id)
-        if not child_records:
-            return []
-        return await self._hydrate_conversations(child_records)
-
-    async def _get_root_record(self, conversation_id: str) -> ConversationRecord:
-        """Resolve the root conversation record without hydrating messages repeatedly."""
-        current = await self.queries.get_conversation(conversation_id)
-        if not current:
-            raise ValueError(f"Conversation {conversation_id} not found")
-
-        while current.parent_conversation_id:
-            parent = await self.queries.get_conversation(str(current.parent_conversation_id))
-            if not parent:
-                break
-            current = parent
-        return current
-
-    async def get_root(self, conversation_id: str) -> Conversation:
-        """Walk up the parent chain to find the root conversation.
-
-        Args:
-            conversation_id: Any conversation ID in the tree
-
-        Returns:
-            The root Conversation object
-
-        Raises:
-            ValueError: If the conversation is not found
-        """
-        root_record = await self._get_root_record(conversation_id)
-        root = await self.get(root_record.conversation_id)
-        if root is None:
-            raise ValueError(f"Conversation {conversation_id} not found")
-        return root
-
-    async def get_session_tree(self, conversation_id: str) -> builtins.list[Conversation]:
-        """Get all conversations in the session tree (root + all descendants).
-
-        Args:
-            conversation_id: Any conversation in the tree
-
-        Returns:
-            List of all Conversation objects in tree (breadth-first order)
-        """
-        root_record = await self._get_root_record(conversation_id)
-
-        tree_records: builtins.list[ConversationRecord] = []
-        queue: builtins.list[ConversationRecord] = [root_record]
-
-        while queue:
-            current = queue.pop(0)
-            tree_records.append(current)
-            children = await self.queries.list_conversations(parent_id=current.conversation_id)
-            queue.extend(children)
-
-        return await self._hydrate_conversations(
-            tree_records,
-            ordered_ids=[record.conversation_id for record in tree_records],
-        )
-
-    async def search_summaries(
-        self,
-        query: str,
-        limit: int = 20,
-        providers: builtins.list[str] | None = None,
-    ) -> builtins.list[ConversationSummary]:
-        """Search conversations and return summaries.
-
-        Args:
-            query: Full-text search query
-            limit: Maximum results
-            providers: Optional list of provider names to filter by
-
-        Returns:
-            List of ConversationSummary objects matching the query
-        """
-        ids, records = await self._search_records(query, limit=limit, providers=providers)
-        if not ids:
-            return []
-        return [conversation_summary_from_record(rec) for rec in records]
-
-    async def search(
-        self,
-        query: str,
-        limit: int = 20,
-        providers: builtins.list[str] | None = None,
-    ) -> builtins.list[Conversation]:
-        """Search conversations using full-text search.
-
-        Args:
-            query: Full-text search query
-            limit: Maximum results
-            providers: Optional list of provider names to filter by
-
-        Returns:
-            List of Conversation objects with all data eager-loaded
-        """
-        ids, records = await self._search_records(query, limit=limit, providers=providers)
-        return await self._hydrate_conversations(records, ordered_ids=ids)
-
-    async def _search_records(
-        self,
-        query: str,
-        *,
-        limit: int,
-        providers: builtins.list[str] | None,
-    ) -> tuple[builtins.list[str], builtins.list[ConversationRecord]]:
-        """Resolve ranked search IDs once and fetch the corresponding records."""
-        ids = await self.queries.search_conversations(query, limit=limit, providers=providers)
-        if not ids:
-            return [], []
-        records = await self.queries.get_conversations_batch(ids)
-        by_id = {record.conversation_id: record for record in records}
-        return ids, [by_id[conversation_id] for conversation_id in ids if conversation_id in by_id]
-
-    async def get_many(self, conversation_ids: builtins.list[str]) -> builtins.list[Conversation]:
-        """Bulk fetch conversations with eager-loaded messages and attachments.
-
-        Uses batch queries to fetch all messages and attachments in 2 queries
-        instead of 2*N individual queries, avoiding connection storms on large
-        databases.
-
-        Args:
-            conversation_ids: List of conversation IDs to fetch
-
-        Returns:
-            List of Conversation objects in the order of input IDs
-        """
-        if not conversation_ids:
-            return []
-
-        records = await self.queries.get_conversations_batch(conversation_ids)
-        return await self._hydrate_conversations(records, ordered_ids=conversation_ids)
-
-    async def iter_messages(
-        self,
-        conversation_id: str,
-        *,
-        dialogue_only: bool = False,
-        limit: int | None = None,
-    ) -> AsyncIterator[Message]:
-        """Stream messages without loading full conversation.
-
-        Memory-efficient iteration for large conversations.
-
-        Args:
-            conversation_id: Conversation ID
-            dialogue_only: If True, only yield user/assistant messages
-            limit: Maximum messages to yield
-
-        Yields:
-            Message objects one at a time
-        """
-        conv_record = await self.queries.get_conversation(conversation_id)
-        provider_name = conv_record.provider_name if conv_record else None
-        async for record in self._backend.iter_messages(
-            conversation_id,
-            dialogue_only=dialogue_only,
-            limit=limit,
-        ):
-            yield message_from_record(record, attachments=[], provider=provider_name)
-
-    async def search_similar(
-        self,
-        text: str,
-        limit: int = 10,
-        vector_provider: VectorProvider | None = None,
-    ) -> builtins.list[Conversation]:
-        """Search by semantic similarity.
-
-        Args:
-            text: Query text
-            limit: Maximum results
-            vector_provider: Vector provider for embedding
-
-        Returns:
-            List of similar Conversation objects
-
-        Raises:
-            ValueError: If no vector provider available
-        """
-        if not vector_provider:
-            raise ValueError("Semantic search requires a vector provider.")
-
-        results = await asyncio.to_thread(
-            vector_provider.query,
-            text,
-            limit=limit * 3,
-        )
-        if not results:
-            return []
-
-        message_ids = [msg_id for msg_id, _ in results]
-        msg_to_conv = await self._get_message_conversation_mapping(message_ids)
-
-        conv_scores: dict[str, float] = {}
-        for msg_id, distance in results:
-            conv_id = msg_to_conv.get(msg_id)
-            if conv_id:
-                # Lower distance = more similar; keep the best (lowest) per conversation
-                conv_scores[conv_id] = min(conv_scores.get(conv_id, float("inf")), distance)
-
-        ranked_ids = sorted(
-            conv_scores.keys(),
-            key=lambda x: conv_scores[x],
-        )[:limit]
-
-        return await self.get_many(ranked_ids)
-
-    async def _get_message_conversation_mapping(
-        self, message_ids: builtins.list[str]
-    ) -> dict[str, str]:
-        """Fetch mapping of message IDs to conversation IDs.
-
-        Args:
-            message_ids: List of message IDs
-
-        Returns:
-            Dict mapping message_id -> conversation_id
-        """
-        if not message_ids:
-            return {}
-
-        placeholders = ",".join("?" * len(message_ids))
-        query = f"SELECT message_id, conversation_id FROM messages WHERE message_id IN ({placeholders})"
-
-        async with self._backend.connection() as conn:
-            cursor = await conn.execute(query, message_ids)
-            rows = await cursor.fetchall()
-
-        return {row["message_id"]: row["conversation_id"] for row in rows}
-
-    def filter(self) -> filters.ConversationFilter:
-        """Create a filter builder for chainable queries.
-
-        Terminal methods (list, first, count, etc.) are async and must be awaited.
-
-        Returns:
-            ConversationFilter instance
-        """
-        from polylogue.lib import filters
-
-        return filters.ConversationFilter(self)
-
-    # --- Write Methods ---
-
-    async def save_conversation(
-        self,
-        conversation: Conversation | ConversationRecord,
-        messages: builtins.list[MessageRecord],
-        attachments: builtins.list[AttachmentRecord],
-        content_blocks: builtins.list[ContentBlockRecord] | None = None,
-    ) -> dict[str, int]:
-        """Save a conversation with its messages and attachments atomically.
-
-        Args:
-            conversation: Conversation model or ConversationRecord to save
-            messages: List of message records
-            attachments: List of attachment records
-
-        Returns:
-            Dictionary with counts:
-                - conversations: New conversations
-                - messages: New messages
-                - attachments: New attachments
-                - skipped_conversations: Unchanged conversations
-                - skipped_messages: Unchanged messages
-                - skipped_attachments: Existing attachments
-        """
-        # Convert Conversation model to ConversationRecord if needed
-        if isinstance(conversation, Conversation):
-            conv_record = self._conversation_to_record(conversation)
-        else:
-            conv_record = conversation
-
-        return await self._save_via_backend(conv_record, messages, attachments, content_blocks or [])
-
-    def _conversation_to_record(self, conversation: Conversation) -> ConversationRecord:
-        """Convert a Conversation model to a ConversationRecord.
-
-        Args:
-            conversation: Conversation model
-
-        Returns:
-            ConversationRecord
-        """
-        from typing import cast
-
-        from polylogue.types import ContentHash, ConversationId
-
-        created_at_str = conversation.created_at.isoformat() if conversation.created_at else None
-        updated_at_str = conversation.updated_at.isoformat() if conversation.updated_at else (created_at_str or None)
-
-        # Try to extract provider_id from canonical id (format: provider:id)
-        # Fallback to whole ID if pattern doesn't match
-        return ConversationRecord(
-            conversation_id=cast(ConversationId, str(conversation.id)),
-            provider_name=conversation.provider,
-            provider_conversation_id=_provider_conversation_id(
-                conversation_id=str(conversation.id),
-                provider=conversation.provider,
-            ),
-            title=conversation.title or "",
-            created_at=created_at_str,
-            updated_at=updated_at_str,
-            content_hash=cast(ContentHash, conversation.metadata.get("content_hash", "")),
-            provider_meta=cast(dict[str, object], conversation.metadata.get("provider_meta", {})),
-            metadata=conversation.metadata,
-        )
-
-    async def _save_via_backend(
-        self,
-        conversation: ConversationRecord,
-        messages: builtins.list[MessageRecord],
-        attachments: builtins.list[AttachmentRecord],
-        content_blocks: builtins.list[ContentBlockRecord] | None = None,
-    ) -> dict[str, int]:
-        """Internal implementation of save_conversation via backend.
-
-        Uses lightweight hash check + UPSERT without redundant full-record
-        reads inside the write lock. The UPSERT ON CONFLICT clauses handle
-        skip detection at the SQL level, so we don't need to pre-fetch
-        existing records.
-
-        Args:
-            conversation: Conversation record
-            messages: List of message records
-            attachments: List of attachment records
-
-        Returns:
-            Dictionary with operation counts
-        """
-        counts: dict[str, int] = {
-            "conversations": 0,
-            "messages": 0,
-            "attachments": 0,
-            "skipped_conversations": 0,
-            "skipped_messages": 0,
-            "skipped_attachments": 0,
-        }
-
-        backend = self._backend
-        if backend is None:
-            raise RuntimeError("Backend is not initialized")
-
-        # Lightweight hash check OUTSIDE the write lock — read-only, no contention
-        existing_hash: str | None = None
-        async with backend.connection() as conn:
-            cursor = await conn.execute(
-                "SELECT content_hash FROM conversations WHERE conversation_id = ?",
-                (conversation.conversation_id,),
-            )
-            row = await cursor.fetchone()
-            if row:
-                existing_hash = row["content_hash"]
-
-        content_unchanged = existing_hash is not None and existing_hash == conversation.content_hash
-
-        # Acquire write lock only for the UPSERT — no pre-fetch reads
-        async with backend.transaction():
-            # Always UPSERT the conversation record (handles title/metadata updates
-            # even when content_hash is unchanged)
-            await backend.save_conversation_record(conversation)
-
-            if content_unchanged:
-                # Content hasn't changed — skip the expensive message/attachment work.
-                # The conversation_content_hash includes all messages, so if the
-                # hash matches, no messages or attachments have been modified.
-                counts["skipped_conversations"] = 1
-                counts["skipped_messages"] = len(messages)
-                counts["skipped_attachments"] = len(attachments)
-            else:
-                counts["conversations"] = 1
-
-                if messages:
-                    # Ensure provider_name is always set on messages (may be empty
-                    # when messages are built outside the prepare pipeline).
-                    pname = conversation.provider_name
-                    if pname:
-                        messages = [
-                            m.model_copy(update={"provider_name": pname})
-                            if not m.provider_name
-                            else m
-                            for m in messages
-                        ]
-                    await backend.save_messages(messages)
-                    counts["messages"] = len(messages)
-                    await backend.upsert_conversation_stats(
-                        conversation.conversation_id,
-                        pname,
-                        messages,
-                    )
-
-                # Collect content blocks from messages if not supplied explicitly
-                all_blocks: builtins.list[ContentBlockRecord] = list(content_blocks or [])
-                for msg in messages:
-                    all_blocks.extend(msg.content_blocks)
-                if all_blocks:
-                    await backend.save_content_blocks(all_blocks)
-
-                new_attachment_ids: set[str] = {str(att.attachment_id) for att in attachments}
-                await backend.prune_attachments(conversation.conversation_id, new_attachment_ids)
-
-                if attachments:
-                    await backend.save_attachments(attachments)
-                    counts["attachments"] = len(attachments)
-
-        invalidate_search_cache()
-        return counts
-
-    async def record_run(self, record: RunRecord) -> None:
-        """Record a pipeline run audit entry.
-
-        Args:
-            record: Run record to persist
-        """
-        await self._backend.record_run(record)
-
-    async def record_publication(self, record: PublicationRecord) -> None:
-        """Persist one publication manifest."""
-        await self._backend.record_publication(record)
-
-    async def get_latest_publication(
-        self,
-        publication_kind: str,
-    ) -> PublicationRecord | None:
-        """Fetch the most recent publication record for one publication kind."""
-        return await self._backend.get_latest_publication(publication_kind)
-
-    # --- Metadata CRUD ---
-
-    async def get_metadata(self, conversation_id: str) -> dict[str, object]:
-        """Get metadata dictionary for a conversation.
-
-        Args:
-            conversation_id: Conversation ID
-
-        Returns:
-            Metadata dictionary (empty if not found)
-        """
-        return await self._backend.get_metadata(conversation_id)
-
-    async def update_metadata(self, conversation_id: str, key: str, value: object) -> None:
-        """Set a single metadata key.
-
-        Args:
-            conversation_id: Conversation ID
-            key: Metadata key
-            value: Value to set
-        """
-        await self._backend.update_metadata(conversation_id, key, value)
-
-    async def delete_metadata(self, conversation_id: str, key: str) -> None:
-        """Remove a metadata key.
-
-        Args:
-            conversation_id: Conversation ID
-            key: Metadata key to delete
-        """
-        await self._backend.delete_metadata(conversation_id, key)
-
-    async def add_tag(self, conversation_id: str, tag: str) -> None:
-        """Add a tag to the conversation's tags list.
-
-        Args:
-            conversation_id: Conversation ID
-            tag: Tag to add
-        """
-        await self._backend.add_tag(conversation_id, tag)
-
-    async def remove_tag(self, conversation_id: str, tag: str) -> None:
-        """Remove a tag from the conversation's tags list.
-
-        Args:
-            conversation_id: Conversation ID
-            tag: Tag to remove
-        """
-        await self._backend.remove_tag(conversation_id, tag)
-
-    async def list_tags(self, *, provider: str | None = None) -> dict[str, int]:
-        """List all tags with counts.
-
-        Args:
-            provider: Optional provider filter
-
-        Returns:
-            Dict of tag -> count, sorted by count descending
-        """
-        return await self._backend.list_tags(provider=provider)
-
-    async def set_metadata(self, conversation_id: str, metadata: dict[str, object]) -> None:
-        """Replace entire metadata dictionary.
-
-        Args:
-            conversation_id: Conversation ID
-            metadata: New metadata dictionary
-        """
-        await self._backend.set_metadata(conversation_id, metadata)
-
-    async def delete_conversation(self, conversation_id: str) -> bool:
-        """Delete conversation and all related records.
-
-        Removes conversation, messages, attachment refs, and FTS index entries.
-
-        Args:
-            conversation_id: Conversation ID to delete
-
-        Returns:
-            True if deleted, False if not found
-        """
-        deleted = await self._backend.delete_conversation(conversation_id)
-        if deleted:
-            invalidate_search_cache()
-        return deleted
-
-    # --- Vector Search Methods ---
-
-    async def embed_conversation(
-        self,
-        conversation_id: str,
-        vector_provider: VectorProvider | None = None,
-    ) -> int:
-        """Generate and store embeddings for a conversation.
-
-        Args:
-            conversation_id: Conversation to embed
-            vector_provider: Optional vector provider (creates default if None)
-
-        Returns:
-            Number of messages embedded
-
-        Raises:
-            ValueError: If no vector provider available
-        """
-        vector_provider = _resolve_optional_vector_provider(vector_provider)
-
-        if vector_provider is None:
-            raise ValueError("No vector provider available. Set VOYAGE_API_KEY.")
-
-        messages = await self.queries.get_messages(conversation_id)
-        if not messages:
-            return 0
-
-        await asyncio.to_thread(
-            vector_provider.upsert,
-            conversation_id,
-            messages,
-        )
-        return len(messages)
-
-    async def similarity_search(
-        self,
-        query: str,
-        limit: int = 10,
-        vector_provider: VectorProvider | None = None,
-    ) -> builtins.list[tuple[str, str, float]]:
-        """Search conversations by semantic similarity.
-
-        Args:
-            query: Search query text
-            limit: Maximum results
-            vector_provider: Optional vector provider
-
-        Returns:
-            List of (conversation_id, message_id, distance) tuples derived from
-            synchronous vector-provider `(message_id, distance)` results
-
-        Raises:
-            ValueError: If no vector provider configured
-        """
-        vector_provider = _resolve_optional_vector_provider(vector_provider)
-
-        if vector_provider is None:
-            raise ValueError("No vector provider configured")
-
-        results = await asyncio.to_thread(
-            vector_provider.query,
-            query,
-            limit=limit,
-        )
-        if not results:
-            return []
-
-        # Batch lookup conversation IDs
-        message_ids = [msg_id for msg_id, _ in results]
-        msg_to_conv = await self._get_message_conversation_mapping(message_ids)
-
-        return [
-            (msg_to_conv[msg_id], msg_id, distance)
-            for msg_id, distance in results
-            if msg_id in msg_to_conv
-        ]
-
-    async def get_archive_stats(self) -> ArchiveStats:
-        """Get comprehensive archive statistics.
-
-        Returns:
-            ArchiveStats with all metrics
-        """
-        from polylogue.lib.stats import ArchiveStats
-
-        async with self._backend.connection() as conn:
-            # Total counts
-            cursor = await conn.execute("SELECT COUNT(*) FROM conversations")
-            conv_count = (await cursor.fetchone())[0]
-
-            cursor = await conn.execute("SELECT COUNT(*) FROM messages")
-            msg_count = (await cursor.fetchone())[0]
-
-            cursor = await conn.execute("SELECT COUNT(*) FROM attachments")
-            att_count = (await cursor.fetchone())[0]
-
-            # Provider breakdown
-            cursor = await conn.execute(
-                """
-                SELECT provider_name, COUNT(*) as count
-                FROM conversations
-                GROUP BY provider_name
-                """
-            )
-            provider_rows = await cursor.fetchall()
-            providers = {row["provider_name"]: row["count"] for row in provider_rows}
-
-            embedding_stats = await read_embedding_stats_async(conn)
-
-        # Get database size
-        db_size = 0
-        try:
-
-            db_size = self._backend.db_path.stat().st_size
-        except Exception as exc:
-            logger.warning("DB size check failed: %s", exc)
-
-        return ArchiveStats(
-            total_conversations=conv_count,
-            total_messages=msg_count,
-            total_attachments=att_count,
-            providers=providers,
-            embedded_conversations=embedding_stats.embedded_conversations,
-            embedded_messages=embedding_stats.embedded_messages,
-            db_size_bytes=db_size,
-        )
-
-
-def _records_to_conversation(
-    conversation: ConversationRecord,
-    messages: list[MessageRecord],
-    attachments: list[AttachmentRecord],
-) -> Conversation:
-    """Convert records to a Conversation model.
-
-    Used by facades and internal migration tools.
-    """
-    return conversation_from_records(conversation, messages, attachments)
-
-
-def _provider_conversation_id(conversation_id: str, provider: str | None) -> str:
-    """Strip only the canonical provider prefix from conversation IDs."""
-    if not provider:
-        return conversation_id
-    prefix = f"{provider}:"
-    return conversation_id[len(prefix) :] if conversation_id.startswith(prefix) else conversation_id
-
-
-def _resolve_optional_vector_provider(
-    vector_provider: VectorProvider | None,
-) -> VectorProvider | None:
-    """Resolve the explicitly supplied provider or create the default one."""
-    if vector_provider is not None:
-        return vector_provider
-
-    from polylogue.storage.search_providers import create_vector_provider
-
-    return create_vector_provider()
 
 
 __all__ = ["ConversationRepository"]
